chore: remove meteor.startup from star-message
