+++ apps/meteor/client/components/message/toolbar/useStarMessageAction.tsx
@@ -1,0 +1,55 @@
+import { useSetting, useToastMessageDispatch } from '@rocket.chat/ui-contexts';
+import { useQueryClient } from '@tanstack/react-query';
+import { useEffect } from 'react';
+
+import { MessageAction } from '../../../../app/ui-utils/client/lib/MessageAction';
+import { sdk } from '../../../../app/utils/client/lib/SDKClient';
+import { toggleStarredMessage } from '../../../lib/mutationEffects/starredMessage';
+import { roomsQueryKeys } from '../../../lib/queryKeys';
+import { roomCoordinator } from '../../../lib/rooms/roomCoordinator';
+
+export const useStarMessageAction = () => {
+	const dispatchToastMessage = useToastMessageDispatch();
+
+	const allowStaring = useSetting('Message_AllowStarring');
+
+	const queryClient = useQueryClient();
+
+	useEffect(() => {
+		MessageAction.addButton({
+			id: 'star-message',
+			icon: 'star',
+			label: 'Star',
+			type: 'interaction',
+			context: ['starred', 'message', 'message-mobile', 'threads', 'federated', 'videoconf', 'videoconf-threads'],
+			async action(_, { message }) {
+				try {
+					await sdk.rest.post('/v1/chat.starMessage', { messageId: message._id });
+					toggleStarredMessage(message, true);
+				} catch (error) {
+					dispatchToastMessage({ type: 'error', message: error });
+				} finally {
+					queryClient.invalidateQueries(roomsQueryKeys.starredMessages(message.rid));
+					queryClient.invalidateQueries(roomsQueryKeys.messageActions(message.rid, message._id));
+				}
+			},
+			condition({ message, subscription, user, room }) {
+				if (subscription == null && allowStaring) {
+					return false;
+				}
+				const isLivechatRoom = roomCoordinator.isLivechatRoom(room.t);
+				if (isLivechatRoom) {
+					return false;
+				}
+
+				return !Array.isArray(message.starred) || !message.starred.find((star: any) => star._id === user?._id);
+			},
+			order: 3,
+			group: 'menu',
+		});
+
+		return () => {
+			MessageAction.removeButton('star-message');
+		};
+	}, [allowStaring, dispatchToastMessage, queryClient]);
+};
