+++ apps/meteor/client/startup/actionButtons/index.ts
@@ -5,6 +5,5 @@ import './jumpToStarMessage';
 import './permalinkPinned';
 import './permalinkStar';
 import './pinMessage';
-import './starMessage';
 import './unpinMessage';
 import './unstarMessage';
